+++ CHANGELOG.md
@@ -6,6 +6,7 @@
 - Fix: No longer updating accounts from deactivated services when manually updating an account.
 - Change: No longer show accounts from deactivated plugins in the player modal window (but still on the user admin 
   page).
+- Update to [Neucore plugin](https://github.com/tkhamez/neucore-plugin) version [next].
 - Small UI improvements.
 
 ## 1.43.0